requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20

# Search Tools
duckduckgo-search>=4.0.0
wikipedia>=1.4.0

#api key and .env linking
python-dotenv
groq